Create integration test class
